Move _build_response() to utils for Django
